Created a new table with possible tags
Created a new table called tags containing id and name of the tag. In the future I will connect the two already existing tables and make it so that companies decide which tags are available, making their own custom ones that might be very specific.

I also removed all of the repeating HTML and put a foreach function to spare the obvious.