Added functionality and chose better naming for .php files as well as titles.
Added a page where students can view the internship in which they showed interest. Also changed some namings for certain files so that they are less confusing as to what they are. I also changed up the main_nav.php and stylesheets.php according to the changes I made. Removed unnecessary comments. Namings are still confusing are still confusing because sometimes I call organization for company and vice versa. index_admin.php is the admin version of view_internships_student.php, so that is confusing as well and namings will be once again improved in the future. Data is being shown in two different ways because in admin view the data to show posts is being shown using php and html, while in student view where the student can see the internships he showed interest in the data is being displayed by using a JS array. This could improve the efficiency and accessability, but might not be worth the effort. Will have to look further into it to decide which method is better.