Fixing the structure of the project as well as adding UX functionality
Moving all success.phps to a seperate folder so it does not infiltrate with the normal pages.

Making so that index2.php is the "real index" for now. It is different in that sense that it includes three files that are located in inc folder. Depending fo the role of the user different .php sites will be displayed in index2.php.

Adding a feedback popup when something is done. It tells you whether action taken by you was successful or not.

Changed some of the file names because they weren't descriptive enough.

Divided CRUD into seperate files because the file was getting big.

removing some auth_checks because they were taking up unneccesary space. Might add them back later.

Added logout button.